Added new models support and cleaned up the classes and structures

+++ src/Models/Pretrained/GPT2PreTrainedModel.php
@@ -14,7 +14,7 @@ class GPT2PreTrainedModel extends PreTrainedModel
 {
     protected int $numHeads;
     protected int $numLayers;
-    protected float $dimKv;
+    protected int $dimKv;
 
     public function __construct(
         AutoConfig               $config,
